LANGCHAIN_OUTPUT_PARSERS/requirements.txt
# LangChain Core
langchain
langchain-core


# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

langchain-google-genai
google-generativeai

langchain-openai